+++ src/Dns/Dns.Test/ScenarioTests/RecordsTests.ps1
@@ -138,30 +138,33 @@ function Test-TrafficManagerProfileRecordSet
 {
 	$zoneName = Get-RandomZoneName
 	$recordName = getAssetname
+	$subscription = getSubscription
 	$resourceGroup = TestSetup-CreateResourceGroup
-	$recordType = "CNAME"
+	$recordType = "A"
 	$zone = $resourceGroup | New-AzDnsZone -Name $zoneName
 
-	# Create a Traffic Manager profile for testing
+	# Synthetic Traffic Manager profile resource Id. Azure DNS does not validate that the
+	# referenced TM profile exists at create time, so we avoid taking a runtime dependency on
+	# the Az.TrafficManager module (which is not referenced by Dns.Test).
 	$tmProfileName = "tmprofile" + $(getAssetname)
-	$tmProfile = New-AzTrafficManagerProfile -Name $tmProfileName -ResourceGroupName $resourceGroup.ResourceGroupName -TrafficRoutingMethod Priority -MonitorProtocol HTTP -MonitorPort 80 -MonitorPath "/" -RelativeDnsName $tmProfileName -Ttl 30
+	$tmProfileId = "/subscriptions/$($subscription)/resourceGroups/$($resourceGroup.ResourceGroupName)/providers/Microsoft.Network/trafficManagerProfiles/$tmProfileName"
 
-	# Create record set with TrafficManagerProfileId (TMLink)
-	$createdRecord = New-AzDnsRecordSet -Name $recordName -ZoneName $zoneName -ResourceGroupName $resourceGroup.ResourceGroupName -RecordType $recordType -TrafficManagerProfileId $tmProfile.Id
+	# TMLink record set (TTL is required for TMLink - service does not auto-derive it like alias)
+	$createdRecord = New-AzDnsRecordSet -Name $recordName -ZoneName $zoneName -ResourceGroupName $resourceGroup.ResourceGroupName -RecordType $recordType -Ttl 3600 -TrafficManagerProfileId $tmProfileId
 
 	Assert-NotNull $createdRecord
 	Assert-AreEqual $zoneName $createdRecord.ZoneName
 	Assert-AreEqual $recordName $createdRecord.Name
 	Assert-AreEqual $resourceGroup.ResourceGroupName $createdRecord.ResourceGroupName
-	Assert-AreEqual $tmProfile.Id $createdRecord.TrafficManagerProfileId
+	Assert-AreEqual $tmProfileId $createdRecord.TrafficManagerProfileId
 
-	# Get and verify read-back
-	$retrievedRecord = Get-AzDnsRecordSet -Name $recordName -ZoneName $zoneName -ResourceGroupName $resourceGroup.ResourceGroupName -RecordType $recordType
-	Assert-AreEqual $tmProfile.Id $retrievedRecord.TrafficManagerProfileId
+	$tmRecord = $zone | Get-AzDnsRecordSet -Name $recordName -RecordType $recordType
+	Assert-AreEqual $tmProfileId $tmRecord.TrafficManagerProfileId
 
-	# Cleanup
-	$retrievedRecord | Remove-AzDnsRecordSet -Confirm:$false
-	Remove-AzTrafficManagerProfile -Name $tmProfileName -ResourceGroupName $resourceGroup.ResourceGroupName -Force
+	$tmRecord | Remove-AzDnsRecordSet
+
+	Assert-ThrowsLike { Get-AzDnsRecordSet -Name $recordName -ZoneName $zoneName -ResourceGroupName $resourceGroup.ResourceGroupName -RecordType $recordType } "*does not exist*"
+
 	Remove-AzDnsZone -Name $zoneName -ResourceGroupName $resourceGroup.ResourceGroupName -Confirm:$false
 	Remove-AzResourceGroup -Name $resourceGroup.ResourceGroupName -Force
 }